Ports + Args
Added the ability to have a single argument to the OS tasks if desired. This allows for numerous tasks that are similar to run the same function with different inputs.

Here is the start of porting over all of the different major microcontroller designs into the RTOS's major structure.
basic Atmega chips should all be ported over. with other major structures having their own def blocks for future implementation.