progress before split
This is where I want to branch the project to remove support for preemption. It does not make sense to support preemption only for Atmega and AtTiny controllers only, so instead, I am going to simplify the project greatly by instead removing support for preemption. This also allows me to fully remove the code from any kind of hardware implementation, thus making it a standalone RTOS.