working coop only
In the advent of stm32, we decided to make ArdRTOS coop only. This build successfully transfers it to be coop only for the vast majority of boards already supported. However, work needs to be done to transfer this over to an Arduino library style instead of a PlatformIO project style.