[0.9] Moving files + Updating comments
This is a major step before releasing as an official Arduino library. getting things into an Arduino library format is just as important as adding an MIT license and professional comments.

MIT licenses were added to every file at the bottom.

The emails in headers were changed to my business email instead of my personal email.